解决跨域

+++ porla_app/app.py
@@ -1,5 +1,5 @@
 from flask import Flask, request, jsonify
-from flask_cors import cross_origin
+from flask_cors import cross_origin, CORS
 import json
 import subprocess
 import random
@@ -12,11 +12,12 @@ import hashlib
 '''
 #	flask服务启动，进行初始化
 app = Flask(__name__)
+CORS(app)
 # databseName = "/..."
-serverPath = "/root/porla/porla/porla/porla/Server/Server"
-clientPath = "/root/porla/porla/porla/porla/Client/Client"
-serverOutputFile = "/root/porla/porla/porla/porla/Server/data_server_output"
-clientOutputFile = "/root/porla/porla/porla/porla/Client/data_client_output"
+serverPath = "/home/lishuai/porla/porla/porla/Server/Server"
+clientPath = "/home/lishuai/porla/porla/porla/Client/Client"
+serverOutputFile = "/data/ls/data_config_file/log/data_server_output"
+clientOutputFile = "/data/ls/data_config_file/log/data_client_output"
 
 databaseName_to_config_and_blockCount = {
     "email_txt": {"config_path": "/data/ls/data_config_file/email_txt", "blockCount": 2097152, "highest_level_index": 21, "destroy_blocks": [], "pre_hash": [], "destroy_hash": []},
@@ -26,9 +27,9 @@ databaseName_to_config_and_blockCount = {
 }
 
 # destroy_block_path = "/home/kt3/porla/porla/porla/Server/destroy_block"
-destroy_log_path = "/root/porla/porla/porla/porla/destroy_log.txt"
+destroy_log_path = "/data/ls/data_config_file/log/destroy_log.txt"
 @app.route('/audit', methods=['POST', 'GET'])
-@cross_origin()
+@cross_origin(origins=["http://10.201.153.203:8001"], methods=["GET", "POST"], allow_headers=["Content-Type"], withCredentials=True)
 def audit():
     if request.method == 'POST':
         data = request.get_json(silent=True)
@@ -42,7 +43,7 @@ def audit():
 
 
 @app.route('/destroy_location', methods=['POST', 'GET'])
-@cross_origin()
+@cross_origin(origins=["http://10.201.153.203:8001"], methods=["GET", "POST"], allow_headers=["Content-Type"], withCredentials=True)
 def destroy_location():
     if request.method == 'POST':
         data = request.get_json(silent=True)
@@ -58,7 +59,7 @@ def destroy_location():
         return " 'it's not a POST operation! "
 
 @app.route('/destroy', methods=['POST', 'GET'])
-@cross_origin()
+@cross_origin(origins=["http://10.201.153.203:8001"], methods=["GET", "POST"], allow_headers=["Content-Type"], withCredentials=True)
 def destroy():
     if request.method == 'POST':
         data = request.get_json(silent=True)
@@ -76,7 +77,7 @@ def destroy():
     
 
 @app.route('/recovery', methods=['POST', 'GET'])
-@cross_origin()
+@cross_origin(origins=["http://10.201.153.203:8001"], methods=["GET", "POST"], allow_headers=["Content-Type"], withCredentials=True)
 def recovery():
     if request.method == 'POST':
         data = request.get_json(silent=True)
@@ -96,6 +97,7 @@ def audit_process(data):
     databaseName = data["databaseName"]
     
     try:
+        serverProcess = None
         serverProcess = subprocess.Popen([serverPath, databaseName], shell=False, stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
         clientProcess = subprocess.Popen([clientPath, databaseName], shell=False, stdout=subprocess.PIPE, stderr=subprocess.STDOUT)
         returncode = clientProcess.wait()
@@ -107,8 +109,9 @@ def audit_process(data):
             "readTime": "-", "computeTime": "-", "preparationTime": "-", "proveTime": "-"
         }
     finally:
-        serverProcess.kill()
-        serverProcess.wait()
+        if serverProcess is not None:
+            serverProcess.kill()
+            serverProcess.wait()
 
     with open(clientOutputFile) as file:
         line = file.readline()
@@ -236,4 +239,4 @@ def calculate_file_hash(file_path, hash_algorithm='sha256'):
     return hash_obj.hexdigest()
 
 if __name__ == '__main__':
-    app.run(host='127.0.0.1', port=7000)
+    app.run(host='127.0.0.1', port=7002)
